+++ src/main/java/br/edu/infnet/application/model/service/NotebookService.java
@@ -1,12 +1,12 @@
 package br.edu.infnet.application.model.service;
 
 import br.edu.infnet.application.model.domain.produtos.Notebook;
-import br.edu.infnet.application.model.domain.Usuario;
 import br.edu.infnet.application.model.repository.NotebookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class NotebookService {
@@ -14,23 +14,37 @@ public class NotebookService {
 	@Autowired
 	private NotebookRepository notebookRepository;
 
+	public Notebook incluir(Notebook notebook) {
+		return notebookRepository.save(notebook);
+	}
+
+	public boolean excluir(Integer id) {
+		Optional<Notebook> byId = notebookRepository.findById(id);
+		if (byId.isPresent()) {
+			notebookRepository.deleteById(id);
+			return true;
+		} else {
+			return false;
+		}
+	}
+
 	public Collection<Notebook> obterLista(){
-		
 		return (Collection<Notebook>) notebookRepository.findAll();
 	}
-	
-	public Collection<Notebook> obterLista(Usuario usuario){
-		
-		return (Collection<Notebook>) notebookRepository.obterLista(usuario.getId());
+
+	public Notebook obterPorId(int codigo) {
+		Optional<Notebook> byId = notebookRepository.findById(codigo);
+		return byId.orElse(null);
 	}
 
-	public void incluir(Notebook notebook) {
+	public Notebook atualizarNotebook(Notebook notebook, int codigo) throws Exception {
+		Optional<Notebook> notebookOptional = notebookRepository.findById(codigo);
+		if (notebookOptional.isPresent()) {
+			notebook.setId(codigo);
+			return notebookRepository.save(notebook);
+		} else {
+			return null;
+		}
 
-		notebookRepository.save(notebook);
 	}
-	
-	public void excluir(Integer id) {
-
-		notebookRepository.deleteById(id);
-	}	
 }
