+++ src/main/java/br/edu/infnet/application/model/service/PedidoService.java
@@ -1,56 +1,50 @@
 package br.edu.infnet.application.model.service;
 
 import br.edu.infnet.application.model.domain.Pedido;
-import br.edu.infnet.application.model.domain.Requerente;
-import br.edu.infnet.application.model.domain.Usuario;
 import br.edu.infnet.application.model.repository.PedidoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PedidoService {
-	
+
 	@Autowired
 	private PedidoRepository pedidoRepository;
 
-	@Autowired
-	private RequerenteService requerenteService;
+	public Pedido incluir(Pedido pedido) {
+		return pedidoRepository.save(pedido);
+	}
 
-	@Autowired
-	private UsuarioService usuarioService;
+	public boolean excluir(Integer id) {
+		Optional<Pedido> byId = pedidoRepository.findById(id);
+		if (byId.isPresent()) {
+			pedidoRepository.deleteById(id);
+			return true;
+		} else {
+			return false;
+		}
+	}
 
 	public Collection<Pedido> obterLista(){
-		
 		return (Collection<Pedido>) pedidoRepository.findAll();
 	}
-	
-	public Collection<Pedido> obterLista(Usuario usuario){
-		
-		List<Pedido> lista = pedidoRepository.obterLista(usuario.getId());
-		
-		if(lista.isEmpty()) {
-			return Collections.emptyList();
+
+	public Pedido obterPorId(int codigo) {
+		Optional<Pedido> byId = pedidoRepository.findById(codigo);
+		return byId.orElse(null);
+	}
+
+	public Pedido atualizarPedido(Pedido pedido, int codigo) throws Exception {
+		Optional<Pedido> pedidoOptional = pedidoRepository.findById(codigo);
+		if (pedidoOptional.isPresent()) {
+			pedido.setId(codigo);
+			return pedidoRepository.save(pedido);
+		} else {
+			return null;
 		}
-		
-		return lista;
-	}
 
-	public void incluir(Pedido pedido) {
-		Requerente incluir = requerenteService.incluir(pedido.getRequerente());
-		Usuario usuario = usuarioService.incluir(new Usuario());
-		pedido.setData(LocalDateTime.now());
-		pedido.setUsuario(usuario);
-		pedido.setRequerente(incluir);
-		pedidoRepository.save(pedido);
 	}
-	
-	public void excluir(Integer id) {
-
-		pedidoRepository.deleteById(id);
-	}	
 }
